+++ frontend/src/components/Modals/EasyModalHeader.tsx
